Handle errors and show error messages as a snackbar, also apply operation to input number
I'm now allowing exceptions to bubble up to the fragment so I can handle them in a user friendly way.

This also slightly changes what an operation applies to which I think makes sense how a real calculator would work. Now you are no longer required to have a number submitted into the stack before doing an operation. The number can still be in the input field. If no number is entered, the operation will apply to the 2 previous on the stack. Also made refinements to DEL and CLEAR with this in mind.